+++ test/rive_native_render_box_test.dart
@@ -34,7 +34,11 @@ base class _FakeRenderTexture extends rive.RenderTexture {
   @override
   Future<ui.Image> toImage() => throw UnimplementedError();
   @override
-  Widget widget({rive.RenderTexturePainter? painter, Key? key}) =>
+  Widget widget({
+    rive.RenderTexturePainter? painter,
+    rive.RenderResolution resolution = const rive.RenderResolution.display(),
+    Key? key,
+  }) =>
       SizedBox.shrink(key: key);
   @override
   void dispose() {}